% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicircle.R
\name{dsemicircle}
\alias{dsemicircle}
\title{Density of the Semicircle Distribution}
\usage{
dsemicircle(x, R, a = 0)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{R}{The radius of the semicircle.}

\item{a}{The shift parameter (default is 0).}
}
\value{
A numeric vector with the density values for each input value of \code{x}.
}
\description{
Computes the density of the semicircle distribution at \code{x} given a radius \code{R} and shift \code{a}.
}
\examples{
dsemicircle(0, R = 1)
dsemicircle(c(0, 0.5, 1), R = 1)
}
\seealso{
\code{\link{semicircle}} for the full documentation on the Wigner semicircle distribution.
}
