% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicircle.R
\name{qsemicircle}
\alias{qsemicircle}
\title{Quantiles of the Semicircle Distribution}
\usage{
qsemicircle(p, R, a = 0)
}
\arguments{
\item{p}{A numeric vector of probabilities.}

\item{R}{The radius of the semicircle.}

\item{a}{The shift parameter (default is 0).}
}
\value{
A numeric vector with the quantiles corresponding to the probabilities \code{p}.
}
\description{
Computes the quantile corresponding to the probability \code{p} for the semicircle distribution.
}
\examples{
qsemicircle(0.5, R = 1)
qsemicircle(c(0.1, 0.5, 0.9), R = 1)
}
\seealso{
\code{\link{semicircle}} for the full documentation on the Wigner semicircle distribution.
}
