% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsemicircle.R
\name{dsemicircle}
\alias{dsemicircle}
\title{Semicircle Distribution Probability Density Function (PDF)}
\usage{
dsemicircle(x, R)
}
\arguments{
\item{x}{A numeric value representing the point at which the PDF is evaluated. It must lie in the range \(\link{-R, R}\).}

\item{R}{A positive numeric value representing the radius parameter of the semicircle distribution.}
}
\value{
A numeric value representing the PDF of the semicircle distribution at point \code{x}.
}
\description{
This function computes the probability density function (PDF) of the semicircle distribution with a given radius \( R \).
}
\details{
The semicircle distribution is defined as:
\deqn{f(x) = \frac{2}{\pi R^2} \sqrt{R^2 - x^2}} for \(-R \leq x \leq R\),
and \deqn{f(x) = 0} for \(|x| > R\).
}
\examples{
dsemicircle(0, 1)   # Should return 2/pi
dsemicircle(0.5, 1) # Should return a positive value
dsemicircle(1, 1)   # Should return 0
dsemicircle(1.5, 1) # Should return 0 (outside the range)

}
