% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psemicircle.R
\name{psemicircle}
\alias{psemicircle}
\title{Semicircle Distribution Cumulative Density Function (CDF)}
\usage{
psemicircle(x, R)
}
\arguments{
\item{x}{A numeric value representing the point at which the CDF is evaluated. It must lie in the range \eqn{[-R, R]}.}

\item{R}{A positive numeric value representing the radius parameter of the semicircle distribution.}
}
\value{
A numeric value representing the CDF of the semicircle distribution at point \code{x}.
}
\description{
This function computes the cumulative density function (CDF) of the semicircle distribution with a given radius \eqn{R}.
}
\examples{
psemicircle(0, 1)   # Should return 0.5
psemicircle(-1, 1)  # Should return 0
psemicircle(1.5, 1) # Should return 0 (outside the range)

}
